+++ src/domains/job/common/components/form/PostContent.jsx
@@ -1,14 +1,32 @@
 import FormItem from '../FormItem'
-const PostContent = ({ register, errors }) => {
+import { Controller } from 'react-hook-form'
+import WriteEditor from '../../../../../components/common/WriteEditor'
+
+const PostContent = ({ control, errors }) => {
   return (
-    <FormItem label='내용' dot={true} register={register} name='text' required={true}>
-      <textarea
-        placeholder='내용을 입력하세요'
-        {...register('text', { required: '내용은 필수입니다' })}
-        className='w-full  h-[360px] border border-dark-gray rounded-md px-4 py-4 focus:outline-none focus:border-main-pink'
+    <FormItem label='내용' dot={true}>
+      <Controller
+        name='text'
+        control={control}
+        rules={{
+          required: '내용은 필수입니다',
+          validate: (value) => {
+            const content = value.getCurrentContent?.()
+            return (content && content.hasText()) || '내용은 필수입니다'
+          },
+        }}
+        render={({ field }) => (
+          <WriteEditor
+            editorState={field.value}
+            onEditorStateChange={field.onChange}
+            placeholder='내용을 입력하세요'
+          />
+        )}
       />
       <div className='flex items-start'>
-        {errors.text && <span className='text-red-500 text-[14px] mt-1'>내용은 필수입니다.</span>}
+        {errors.text && (
+          <span className='text-red-500 text-[14px] mt-1'>{errors.text.message}</span>
+        )}
       </div>
     </FormItem>
   )
